feat: 장바구니 기능 구현

+++ src/main/java/com/delivery/yunyun/dto/response/StoreMenuListResponse.java
@@ -2,10 +2,12 @@ package com.delivery.yunyun.dto.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record StoreMenuListResponse(
         String menuName,
-        Integer price,
+        BigDecimal price,
         String introduction
 ) {
 }
